+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixed $select not properly working on unitOfMeasurement sub-properties.
 * Fixed #1934, incorrect URL parameter in docs.
 * Changed header handling to allow for duplicate headers.
+* Fixed MultiDatastream not picking up ID type of Thing and Sensor.
 
 
 ## Release version 2.3.0